Fix date conversion problem when creating bet
Now the chosen expiry date is converted to the server time instead of
basing it on the user local time. This fixes the problem where a bet
would be considered to be "in the past" if the user local time and given
bet_expiry_date were lower than the current server time.